Added weapon damage amounts
Still fighting the bug on weapon button click. Came up with a possible solution. Once player clicks a button, hide the BattleForm and show a new Cutscene userform that will show while the code is running. This would also allow for player customization based on weapon type selected to use.